Replace `FeedPresenter` from iOS module with `FeedPresenter` from shared module
