+++ EssentialFeed/Feed Presentation/FeedLoadingViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  FeedLoadingViewModel.swift
+//  EssentialFeed
+//
+//  Created by Macbook Pro on 22/9/23.
+//
+
+import Foundation
+
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
