+++ EssentialFeed/Feed Presentation/FeedViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  FeedViewModel.swift
+//  EssentialFeed
+//
+//  Created by Macbook Pro on 22/9/23.
+//
+
+import Foundation
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
