--- EssentialFeediOS/Feed Presentation/el.lproj/Feed.strings
@@ -1,9 +1,0 @@
-/* 
-  Feed.strings
-  EssentialFeed
-
-  Created by Macbook Pro on 20/9/23.
-  
-*/
-"FEED_VIEW_TITLE" = "To Feed ou";
-"FEED_VIEW_CONNECTION_ERROR" = "Δεν ήταν δυνατή η σύνδεση στο διακομιστή";
