--- EssentialFeediOS/Feed Presentation/en.lproj/Feed.strings
@@ -1,9 +1,0 @@
-/* 
-  Feed.strings
-  EssentialFeed
-
-  Created by Macbook Pro on 20/9/23.
-  
-*/
-"FEED_VIEW_TITLE" = "My Feed";
-"FEED_VIEW_CONNECTION_ERROR" = "Couldn't connect to server";
