--- EssentialFeediOS/Feed Presentation/pt-BR.lproj/Feed.strings
@@ -1,10 +1,0 @@
-/* 
-  Feed.strings
-  EssentialFeed
-
-  Created by Macbook Pro on 20/9/23.
-  
-*/
-"FEED_VIEW_TITLE" = "Meu Feed";
-"FEED_VIEW_CONNECTION_ERROR" = "Não foi possível conectar com o servidor";
-
