+++ EssentialFeediOS/Feed UI/Composers/FeedViewAdapter.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import UIKit
+import EssentialFeed
 
 final class FeedViewAdapter: FeedView {
     private weak var controller: FeedViewController?
