+++ EssentialFeediOSTests/Feed Presentation/FeedLocalizationTests.swift
@@ -6,7 +6,7 @@
 //
 
 import XCTest
-@testable import EssentialFeediOS
+import EssentialFeed
 
 final class FeedLocalizationTests: XCTestCase {
