+++ EssentialFeediOSTests/Feed UI/Helpers/FeedViewController+Localization.swift
@@ -6,13 +6,13 @@
 //
 
 import Foundation
-import EssentialFeediOS
+import EssentialFeed
 import XCTest
 
 extension FeedUIIntegrationTests {
     func localized(_ key: String, file: StaticString = #file, line: UInt = #line) -> String {
         let table = "Feed"
-        let bundle = Bundle(for: FeedViewController.self)
+        let bundle = Bundle(for: FeedPresenter.self)
         let value = bundle.localizedString(forKey: key, value: nil, table: table)
         if value == key {
             XCTFail("Missing localized string for key: \(key) in table: \(table)", file: file, line: line)
